refactor(fe-settings): extract sub-components to reduce complexity
Reduced max-lines and complexity warnings in settings pages:
- PendingSkillsSection: extracted SKILL_MARKDOWN_COMPONENTS + usePendingSkillActions hook (181→98 lines); removed unused _node var
- PublicDomainSection: extracted PublicDomainHeader/Body + result sub-components (189→141 lines, dropped 2 complexity warnings)
- SettingsContext: extracted useNodeHealthCheck hook for node health checking (6→0 warnings)
- BackupsSettingsPage: extracted state management into useBackupState hook for 1488→1404 lines

No behaviour changes. vitest --changed green.

Closes #lint-fe-settings
